+++ Lorr/Engine/Resource/IResource.hh
@@ -28,7 +28,7 @@ namespace Lorr
         }
 
     protected:
-        Identifier m_Ident = EmptyIdentifier;
+        Identifier m_Ident = kEmptyIdentifier;
     };
 
 }  // namespace Lorr